requirements.txt
APScheduler==3.6.3
astroid==2.5.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
DateTime==4.3
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.7.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-telegram-bot==13.3
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.3
wrapt==1.12.1
zope.interface==5.2.0
